Move simple-p2p.cc to DataRate; fix small bug in DataRate and add include guards

+++ src/common/data-rate.cc
@@ -80,7 +80,7 @@ uint64_t DataRate::Parse(const std::string s)
 
 double DataRate::CalculateTxTime(uint32_t bytes) const
 {
-  return bytes*8/m_bps;
+  return static_cast<double>(bytes)*8/m_bps;
 }
 
 uint64_t DataRate::GetBitRate() const
